requirements.txt
pytorch
tqdm
nltk
sklearn
transformers